refactor out export code
