+++ src/test/java/bio/terra/service/snapshot/SnapshotIntegrationTest.java
@@ -23,7 +23,6 @@ import bio.terra.model.SnapshotRequestModel;
 import bio.terra.model.SnapshotSummaryModel;
 import bio.terra.service.iam.IamResourceType;
 import bio.terra.service.iam.IamRole;
-import bio.terra.service.tabulardata.google.bigquery.BigQueryPdao;
 import com.google.cloud.bigquery.BigQuery;
 import com.google.cloud.bigquery.TableResult;
 import java.util.ArrayList;
@@ -59,7 +58,6 @@ public class SnapshotIntegrationTest extends UsersBase {
   @Autowired private DataRepoFixtures dataRepoFixtures;
 
   @Autowired private AuthService authService;
-  @Autowired private BigQueryPdao bigQueryPdao;
 
   private static final Logger logger = LoggerFactory.getLogger(SnapshotIntegrationTest.class);
   private UUID profileId;
